Crate eth_types

Ethereum and Evm types used to deserialize responses from web3 / geth.

Re-exports

pub use keccak::keccak256;

pub use keccak::Keccak;

pub use bytecode::Bytecode;

pub use error::Error;

Modules

bytecode

EVM byte code generator

error

Error module for the eth-types crate

evm_types

Evm types needed for parsing instruction sets as well

geth_types

Types needed for generating Ethereum traces

keccak

Plain keccak256 implementation

macros

Collection of utility macros used within this crate.

sign_types

secp256k1 signature types and helper functions.

Macros

address

Create an Address from a hex string. Panics on invalid input.

bytecode

EVM code macro

evm_unimplemented

Triggers a signal that a feature is unimplemented, panics if warn-unimplemented feature is not set

word

Create a Word from a hex string. Panics on invalid input.

word_map

Create a Word to Word HashMap from pairs of hex strings. Panics on invalid input.

Structs

AccessList

Access list

Block

The block type returned from RPC calls.

Bytes

Wrapper type around Bytes to deserialize/serialize “0x” prefixed ethereum hex strings

DebugU256

Little-endian large integer type 256-bit unsigned integer.

EIP1186ProofResponse

Struct used to define the result of eth_getProof call

GethExecTrace

The execution trace type returned by geth RPC debug_trace* methods. Corresponds to ExecutionResult in go-ethereum/internal/ethapi/api.go. The deserialization truncates the memory of each step in struct_logs to the memory size before the expansion, so that it corresponds to the memory before the step is executed.

H64

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

Signature

An ECDSA signature

StorageProof

Struct used to define the storage proof

Transaction

Details of a signed transaction

U64

Little-endian large integer type Unsigned 64-bit integer.

U256

Little-endian large integer type 256-bit unsigned integer.

U512

Little-endian large integer type 512-bits unsigned integer.

Traits

BigEndianHash

Field

Trait used to reduce verbosity with the declaration of the [PrimeField] trait and its repr.

ToAddress

Trait used to convert a type to a Address.

ToBigEndian

Trait uset do convert a scalar value to a 32 byte array in big endian.

ToLittleEndian

Trait used to convert a scalar value to a 32 byte array in little endian.

ToScalar

Trait used to define types that can be converted to a 256 bit scalar value.

ToWord

Trait used to convert a type to a Word.

Type Definitions

Address

Hash

Ethereum Hash (256 bits).

Word

Ethereum Word (256 bits).