Module eth_types::evm_types

Evm types needed for parsing instruction sets as well

Re-exports

pub use memory::Memory;

pub use memory::MemoryAddress;

pub use opcode_ids::OpcodeId;

pub use stack::Stack;

pub use stack::StackAddress;

pub use storage::Storage;

Modules

gas_utils

Utility functions to help calculate gas

memory

Doc this

opcode_ids

Doc this

stack

Doc this

storage

Doc this

Structs

GasCost

Defines the gas consumption.

Constants

GAS_STIPEND_CALL_WITH_VALUE

Gas stipend when CALL or CALLCODE is attached with value.

INIT_CODE_WORD_GAS

Once per word of the init code when creating a contract.

INVALID_INIT_CODE_FIRST_BYTE

According to EIP-3541, disallow new code starting with 0xEF to be deployed.

MAX_EXPANDED_MEMORY_ADDRESS

This constant ((2^32 - 1) * 32) is the highest number that can be used without overflowing the square operation of gas calculation. https://github.com/ethereum/go-ethereum/blob/e6b6a8b738069ad0579f6798ee59fde93ed13b43/core/vm/gas_table.go#L38

MAX_REFUND_QUOTIENT_OF_GAS_USED

Quotient for max refund of gas used