Struct eth_types::evm_types::memory::Memory

pub struct Memory(pub Vec<u8>);

Represents a snapshot of the EVM memory state at a certain execution step height.

Tuple Fields
0: Vec<u8>
Implementations

impl Memory

pub const fn new() -> Memory

Generate an new empty instance of EVM memory.

pub fn is_empty(&self) -> bool

Returns true if memory contains no elements.

pub fn len(&self) -> usize

Returns the number of elements in the memory.

pub fn push<T: AsRef<[u8]>>(&mut self, input: T)

Pushes a set of bytes or an Word in the last Memory position.

pub fn last_filled_addr(&self) -> MemoryAddress

Returns the last memory address written at this execution step height.

pub fn read_word(&self, addr: MemoryAddress) -> Word

Reads an entire Word which starts at the provided MemoryAddress addr and finnishes at addr + 32.

pub fn read_chunk( &self, offset: MemoryAddress, length: MemoryAddress ) -> Vec<u8>

Reads an chunk of memory[offset..offset+length]. Zeros will be padded if index out of range.

pub fn word_size(&self) -> usize

Returns the size of memory in word.

pub fn extend_at_least(&mut self, minimal_length: usize)

Resize the memory for at least length and align to 32 bytes.

pub fn copy_from( &mut self, dst_offset: Word, src_offset: Word, length: Word, data: &[u8] )

Copy source data to memory. Used in (ext)codecopy/calldatacopy.